Expose the OvernightIndexedCouponPricer
